[REF] Move .utils scripts to .stats
