+++ brainstat/stats/SurfStatResels.py
@@ -2,7 +2,7 @@ import numpy as np
 from numpy import concatenate as cat
 from scipy.linalg import toeplitz
 from scipy.sparse import csr_matrix
-from brainstat.unassigned.matlab_functions import row_ismember, interp1
+from brainstat.stats.matlab_functions import row_ismember, interp1
 from brainstat.stats.SurfStatEdg import SurfStatEdg
