+++ brainstat/stats/stat_threshold.py
@@ -1,7 +1,7 @@
 import math
 import numpy as np
 from scipy.special import betaln, gammaln, gamma
-from brainstat.unassigned.matlab_functions import interp1, colon
+from brainstat.stats.matlab_functions import interp1, colon
 
 
 def gammalni(n):
